+++ actionview/test/fixtures/mascot.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class Mascot < ActiveRecord::Base
-  belongs_to :company
+  belongs_to :company, optional: false
 end